requirements.txt
python-dotenv
py-cord
pandas
numpy
openpyxl
bs4
playwright
pytz